requirements.txt
streamlit 
opencv-python
numpy
pandas 
matplotlib
pillow